+++ src/UDA/Query/Insert.php
@@ -23,7 +23,7 @@ use UDA\Query\Dialect\InsertState;
 /**
  * INSERT query builder that produces Sql objects for execution
  */
-final class Insert extends Builder
+final class Insert extends \UDA\Query
 {
     use ConsumesCtes;
     /** @var ?string Table to insert into */